+++ CHANGELOG.md
@@ -40,8 +40,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMPullDistributionPoint Resource
 - Added ConvertTo-AnyCimInstance to the ResourceHelper
 - Added CMSiteMaintenance Resource
-- Added CMDistributionGroup Resource
+- Added CMAdministrativeUser Resource
 - Added Compare-MultipleCompares to the ResourceHelper
+- Added CMDistributionGroup Resource
 
 ### Changed